feat: add hooks.utils
